+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ## v1.24.0
 
-*2025-08-13*
+*2025-08-14*
 
 **Added**
 
 - [`AstNode.prototype.remove`](https://github.com/bhsd-harry/wikiparser-node/wiki/AstNode-%28EN%29#remove) now takes an optional `ownLine` parameter to remove the current line if it is empty
 - Quick-fix suggestions for the [nested-link](https://github.com/bhsd-harry/wikiparser-node/wiki/nested-link) rule
 - [`LanguageService.prototype.provideRefactoringAction`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService-%28EN%29#providerefactoringaction) that provides refactoring actions for the selected text
+- [`LanguageService.prototype.resolveCodeAction`](https://github.com/bhsd-harry/wikiparser-node/wiki/LanguageService-%28EN%29#resolvecodeaction) that resolves fix-all code actions
 
 **Fixed**